feat: support asynchronous custom filters
BREAKING CHANGE: custom filters do not anymore return the value
associated to the key of the custom filter. Instead the key is
remove from the filters and the returned object populates the
filters. In order to keep the same behavior for an existing
custom filter, the following modification must be implemented:

```diff
crud('/admin/users', sequelizeCrud(User), {
  filters: {
    email: value => ({
-      [Op.iLike]: value,
+      email: {[Op.iLike]: value },
      },
    }),
  },
})
```

Asynchronous custom filters can be implemented:

```ts
const filters = {
  category: async value => {
    const posts = await Post.find({ category: value })

    return {
      id: posts.map(post => post.userId),
    }
  },
}
```

The category key will be removed and the id key will be added
with the value computed by the custom filter handler