feat: support asynchronous custom filters
BREAKING CHANGE: custom filter handlers do not anymore return the value
associated to the key of the given handler. Instead the returned value must be an object that will be merged in the filters passed to the getter handler. In order to keep the same behavior for an existing
custom filter, the following modification must be implemented:

```diff
crud('/admin/users', sequelizeCrud(User), {
  filters: {
    email: value => ({
-      [Op.iLike]: value,
+      email: {[Op.iLike]: value },
      },
    }),
  },
})
```

Asynchronous custom filters are now supported:

```ts
const filters = {
  categoryName: async value => {
    const category = await Category.findOne({ name: value }).orFail()

    // this object will be merged in filters given to the getter handler
    return {
      categoryId: category.id,
    }
  },
}
```